Add flash kind for successful interactions
Introduce a `:success` kind for flash messages in the generator
output and update relevant documentation.

Success messages that previously used `:info` now use `:success`,
ensuring they continue to appear in green with a "Success!" title.

The icon for success messages is now a checkmark.

Info messages do now appear in blue with the title "Info".

In certain cases (e.g. password reset), `:info` is retained to avoid
disclosing too much information (such as the existence of an account).

Rationale for those changes:

- Differentiate informational messages from successful actions

- Avoid confusion from informational messages styled in green, which
  can imply success

- Ensure `:info` is used for neutral messages, and use `:success` to
  clearly indicate completed actions